Read animation keys and separate Roi from Model
There is global data for every model which contains one Roi and maybe
multiple children Rois. The position of the child Rois are not relative
to the parent Roi. Instead they are stored in the translation key of the
global animation data of the model.

It also reads the other animation data now and stores them all in an
animation Node.